create 2nd spec from hbx enrollment
